Actually, I need to set DH_BITS...
It turns out that I do need to set DH_BITS. Seeing as DH paramaters are
no longer generated (see RFC7919), there is little value in using smaller
ones. I was just going to hardcode it to 8192 (the largest), but there are
other valid options. It doesn't really make sense to make it easily adjustable
(via a switch, but I decided not to hardcode it and have left an env var
if customisation is desired. FWIW DH is only used for older legacy ciphers
these days anyway, TLS v1.3 doesn't even use them...